+++ mlflow_cratedb/server.py
@@ -1,2 +1,8 @@
 # Intercept server entrypoint for monkeypatching.
-from mlflow.server import app  # noqa: F401
+# ruff: noqa: ERA001
+
+# Use Flask with gunicorn
+# from mlflow.server import app  # noqa: F401
+
+# Use FastAPI with uvicorn
+from mlflow.server.fastapi_app import app  # noqa: F401