+++ CHANGES.md
@@ -4,6 +4,7 @@
 ## in progress
 - Amalgamation: Started supporting huey jobs that are invoked in subprocesses
 - Model: Added tables `jobs`, `online_scoring_configs`, `scorers`, `scorer_versions`
+- Server: Switched from `gunicorn` to `uvicorn` as suggested
 
 ## 2026-03-29 v3.10.1
 - Updated to [MLflow 3.10.1]